api/pyproject.toml
[project]
name = "frame-extractor"
version = "0.1.0"
description = "Video frame extractor API - extracts unique frames via SSIM similarity"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "av>=12.0.0",
    "scikit-image>=0.23.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.6",
    "openai>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
